backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.3
netCDF4==1.6.5
xarray==2024.1.0
dask[complete]==2024.1.0

# HTTP requests
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.3

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.2.0

# LLM integration
ollama==0.1.6
openai==1.12.0
google-generativeai==0.3.2

# RAG / Vector Database
qdrant-client==1.7.3
sentence-transformers==2.3.1

# Ocean science calculations
gsw==3.6.17

# CLI tools
click==8.2.1
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2

# Testing
pytest==8.3.4

